doc-analyzer：所有Handle从 interface+singleton 改成 singleton 模式
