+++ doc-analyzer/src/main/java/com/spldeolin/allison1875/docanalyzer/DocAnalyzerModule.java
@@ -7,17 +7,10 @@ import com.spldeolin.allison1875.base.ancestor.Allison1875Module;
 import com.spldeolin.allison1875.docanalyzer.handle.AccessDescriptionHandle;
 import com.spldeolin.allison1875.docanalyzer.handle.AnalyzeCustomValidationHandle;
 import com.spldeolin.allison1875.docanalyzer.handle.AnalyzeEnumConstantHandle;
-import com.spldeolin.allison1875.docanalyzer.handle.DefaultMoreHandlerAnalysisHandle;
 import com.spldeolin.allison1875.docanalyzer.handle.MoreHandlerAnalysisHandle;
 import com.spldeolin.allison1875.docanalyzer.handle.MoreJpdvAnalysisHandle;
 import com.spldeolin.allison1875.docanalyzer.handle.ObtainConcernedResponseBodyHandle;
 import com.spldeolin.allison1875.docanalyzer.handle.SpecificFieldDescriptionsHandle;
-import com.spldeolin.allison1875.docanalyzer.handle.impl.DefaultAnalyzeCustomValidationHandle;
-import com.spldeolin.allison1875.docanalyzer.handle.impl.DefaultAnalyzeEnumConstantHandle;
-import com.spldeolin.allison1875.docanalyzer.handle.impl.DefaultMoreJpdvAnalysisHandle;
-import com.spldeolin.allison1875.docanalyzer.handle.impl.DefaultObtainConcernedResponseBodyHandle;
-import com.spldeolin.allison1875.docanalyzer.handle.impl.DefaultSpecificFieldDescriptionsHandle;
-import com.spldeolin.allison1875.docanalyzer.handleimpl.DefaultAccessDescriptionHandle;
 import com.spldeolin.allison1875.docanalyzer.processor.DocAnalyzer;
 import lombok.extern.log4j.Log4j2;
 
@@ -29,13 +22,13 @@ public class DocAnalyzerModule extends Allison1875Module {
 
     @Override
     protected void configure() {
-        bind(AnalyzeCustomValidationHandle.class).to(DefaultAnalyzeCustomValidationHandle.class);
-        bind(AnalyzeEnumConstantHandle.class).to(DefaultAnalyzeEnumConstantHandle.class);
-        bind(MoreJpdvAnalysisHandle.class).to(DefaultMoreJpdvAnalysisHandle.class);
-        bind(ObtainConcernedResponseBodyHandle.class).to(DefaultObtainConcernedResponseBodyHandle.class);
-        bind(SpecificFieldDescriptionsHandle.class).to(DefaultSpecificFieldDescriptionsHandle.class);
-        bind(AccessDescriptionHandle.class).to(DefaultAccessDescriptionHandle.class);
-        bind(MoreHandlerAnalysisHandle.class).to(DefaultMoreHandlerAnalysisHandle.class);
+        bind(AnalyzeCustomValidationHandle.class).to(AnalyzeCustomValidationHandle.class);
+        bind(AnalyzeEnumConstantHandle.class).to(AnalyzeEnumConstantHandle.class);
+        bind(MoreJpdvAnalysisHandle.class).to(MoreJpdvAnalysisHandle.class);
+        bind(ObtainConcernedResponseBodyHandle.class).to(ObtainConcernedResponseBodyHandle.class);
+        bind(SpecificFieldDescriptionsHandle.class).to(SpecificFieldDescriptionsHandle.class);
+        bind(AccessDescriptionHandle.class).to(AccessDescriptionHandle.class);
+        bind(MoreHandlerAnalysisHandle.class).to(MoreHandlerAnalysisHandle.class);
     }
 
     @Override
