+++ doc-analyzer/src/main/java/com/spldeolin/allison1875/docanalyzer/handle/AccessDescriptionHandle.java
@@ -2,15 +2,22 @@ package com.spldeolin.allison1875.docanalyzer.handle;
 
 import java.util.Collection;
 import com.github.javaparser.ast.body.FieldDeclaration;
+import com.google.inject.Singleton;
+import com.spldeolin.allison1875.base.util.ast.JavadocDescriptions;
 import com.spldeolin.allison1875.docanalyzer.javabean.HandlerFullDto;
 
 /**
  * @author Deolin 2021-04-06
  */
-public interface AccessDescriptionHandle {
+@Singleton
+public class AccessDescriptionHandle {
 
-    Collection<String> accessMethod(HandlerFullDto handlerFullDto);
+    public Collection<String> accessMethod(HandlerFullDto handlerFullDto) {
+        return JavadocDescriptions.getAsLines(handlerFullDto.getMd());
+    }
 
-    Collection<String> accessField(FieldDeclaration field);
+    public Collection<String> accessField(FieldDeclaration field) {
+        return JavadocDescriptions.getAsLines(field);
+    }
 
 }
