+++ doc-analyzer/src/main/java/com/spldeolin/allison1875/docanalyzer/handle/AnalyzeCustomValidationHandle.java
@@ -2,15 +2,18 @@ package com.spldeolin.allison1875.docanalyzer.handle;
 
 import java.lang.reflect.AnnotatedElement;
 import java.util.Collection;
+import javax.inject.Singleton;
+import com.google.common.collect.Lists;
 import com.spldeolin.allison1875.docanalyzer.javabean.ValidatorDto;
 
 /**
- * 解析自定义校验注解策略
- *
  * @author Deolin 2020-06-18
  */
-public interface AnalyzeCustomValidationHandle {
+@Singleton
+public class AnalyzeCustomValidationHandle {
 
-    Collection<ValidatorDto> analyzeCustomValidation(AnnotatedElement annotatedElement);
+    public Collection<ValidatorDto> analyzeCustomValidation(AnnotatedElement annotatedElement) {
+        return Lists.newArrayList();
+    }
 
 }
