+++ doc-analyzer/src/main/java/com/spldeolin/allison1875/docanalyzer/handle/SpecificFieldDescriptionsHandle.java
@@ -1,17 +1,17 @@
 package com.spldeolin.allison1875.docanalyzer.handle;
 
+import javax.inject.Singleton;
+import com.google.common.collect.HashBasedTable;
 import com.google.common.collect.Table;
 
 /**
- * 提供 对特定Field的描述 的策略
- *
  * @author Deolin 2020-07-08
  */
-public interface SpecificFieldDescriptionsHandle {
+@Singleton
+public class SpecificFieldDescriptionsHandle {
 
-    /**
-     * @return 类的全限定名，字段名，描述
-     */
-    Table<String, String, String> provideSpecificFieldDescriptions();
+    public Table<String, String, String> provideSpecificFieldDescriptions() {
+        return HashBasedTable.create();
+    }
 
 }
