--- doc-analyzer/src/main/java/com/spldeolin/allison1875/docanalyzer/handle/impl/DefaultMoreJpdvAnalysisHandle.java
@@ -1,22 +1,0 @@
-package com.spldeolin.allison1875.docanalyzer.handle.impl;
-
-import java.lang.reflect.Field;
-import com.spldeolin.allison1875.docanalyzer.DocAnalyzerConfig;
-import com.spldeolin.allison1875.docanalyzer.handle.MoreJpdvAnalysisHandle;
-
-/**
- * @author Deolin 2020-12-02
- */
-public class DefaultMoreJpdvAnalysisHandle implements MoreJpdvAnalysisHandle {
-
-    @Override
-    public Object moreAnalysisFromField(Field field) {
-        return null;
-    }
-
-    @Override
-    public String moreJpdvToString(Object dto, DocAnalyzerConfig docAnalyzerConfig) {
-        return null;
-    }
-
-}
